src/main/java/zeroj/zerojspring/repository/MemoryMemberRepository.java
package zeroj.zerojspring.repository;

import zeroj.zerojspring.domain.Member;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;

    @Override
    public Member save(Member member) {
        //세이브하기전 이름은 넣어놓은 상태. Id의 경우 시스템에 저장할때 시스템이 지정해주는 것
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByName(String name) {
        //루프가 돌면서 Map에서 하나 찾으면 걔를 반환. 없으면 Optional에 Null이 포함되서 반환
        return store.values().stream()
                .filter(member -> member.getName().equals(name))
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore(){
        store.clear();
    }
}
